requirements.txt
Flask
psycopg2-binary
pytest
pytest-flask
pytest-cov
